scripts/social_drafts/requirements.txt
python-dotenv==1.0.1
openai==1.59.8
httpx==0.28.1
